feat: Make player's email field nullable in forms and update tests (#8)
In this commit, the player's email field has been made nullable in the forms, removing the requirement for it to be a mandatory field. The tests have been updated accordingly to reflect this change.

By making the email field nullable, users are no longer required to provide an email address when submitting the form. This allows for more flexibility in the data entry process.